jamboree-vouchers/requirements.txt
appdirs==1.4.3
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
six==1.10.0
